+++ python/GafferSceneTest/ShaderTweaksTest.py
@@ -179,12 +179,15 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		tweaks["filter"].setInput( planeFilter["out"] )
 		tweaks["shader"].setValue( "surface" )
 
+		# Insert a connection from the new `textureShader`, and
+		# check it worked.
+
 		tweaks["tweaks"].addChild( Gaffer.TweakPlug( "c", Gaffer.Color3fPlug() ) )
-		tweaks["tweaks"][0]["value"].setInput( textureShader["out"] )
+		tweaks["tweaks"][0]["value"].setInput( textureShader["out"]["c"] )
 
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 2 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture", "out" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture", "c" ) )
 
 		tweakedNetwork.removeShader( "texture" )
 		self.assertEqual( tweakedNetwork, originalNetwork )
@@ -193,9 +196,36 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( tweakedNetwork.getShader( "texture" ).parameters["c"].value, imath.Color3f( 1, 2, 3 ) )
 
+		# We can't use numeric tweaks modes when inserting a connection, so
+		# check that trying causes an error.
+
 		tweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Multiply )
-		with self.assertRaisesRegex( RuntimeError, "Mode must be \"Replace\" when inserting a connection" ) :
+		with self.assertRaisesRegex( RuntimeError, 'Mode must be "Replace" or "Create" when inserting a connection' ) :
 			tweaks["out"].attributes( "/plane" )
+
+		# Try to make a connection to a parameter that doesn't exist. In Replace
+		# mode this should error.
+
+		tweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Replace )
+		tweaks["tweaks"][0]["name"].setValue( "newParameter" )
+
+		with self.assertRaisesRegex( RuntimeError, 'ShaderTweaks.out.attributes : Cannot apply tweak "newParameter" because shader "surface" does not have parameter "newParameter"' ) :
+			tweaks["out"].attributes( "/plane" )
+
+		# But it should be allowed in Create mode.
+
+		tweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Create )
+		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
+		self.assertEqual( len( tweakedNetwork ), 2 )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "newParameter" ) ), ( "texture", "c" ) )
+
+		# And Create mode should also be allowed to overwrite an existing
+		# connection, just as it would overwrite an existing value.
+
+		upstreamTextureShader = GafferSceneTest.TestShader( "upstreamTexture" )
+		shader["parameters"]["c"].setInput( upstreamTextureShader["out"]["c"] )
+		tweaks["tweaks"][0]["name"].setValue( "c" )
+		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 
 	def testConnectSpecificOutputParameter( self ) :
 
@@ -245,7 +275,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		shader["type"].setValue( "surface" )
 
 		textureShader1 = GafferSceneTest.TestShader( "texture1" )
-		shader["parameters"]["c"].setInput( textureShader1["out"] )
+		shader["parameters"]["c"].setInput( textureShader1["out"]["c"] )
 
 		planeFilter = GafferScene.PathFilter()
 		planeFilter["paths"].setValue( IECore.StringVectorData( [ "/plane" ] ) )
@@ -257,7 +287,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 
 		originalNetwork = assignment["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( originalNetwork ), 2 )
-		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture1", "out" ) )
+		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture1", "c" ) )
 
 		textureShader2 = GafferSceneTest.TestShader( "texture2" )
 
@@ -267,11 +297,11 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		tweaks["shader"].setValue( "surface" )
 
 		tweaks["tweaks"].addChild( Gaffer.TweakPlug( "c", Gaffer.Color3fPlug() ) )
-		tweaks["tweaks"][0]["value"].setInput( textureShader2["out"] )
+		tweaks["tweaks"][0]["value"].setInput( textureShader2["out"]["c"] )
 
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
 		self.assertEqual( len( tweakedNetwork ), 2 )
-		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture2", "out" ) )
+		self.assertEqual( tweakedNetwork.input( ( "surface", "c" ) ), ( "texture2", "c" ) )
 
 		textureShader2["enabled"].setValue( False )
 		tweakedNetwork = tweaks["out"].attributes( "/plane" )["surface"]
@@ -289,7 +319,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		shader["type"].setValue( "surface" )
 
 		textureShader = GafferSceneTest.TestShader( "texture1" )
-		shader["parameters"]["c"].setInput( textureShader["out"] )
+		shader["parameters"]["c"].setInput( textureShader["out"]["c"] )
 
 		planeFilter = GafferScene.PathFilter()
 		planeFilter["paths"].setValue( IECore.StringVectorData( [ "/plane" ] ) )
@@ -308,11 +338,11 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 
 		for mode in Gaffer.TweakPlug.Mode.values :
 
-			if mode == Gaffer.TweakPlug.Mode.Replace :
+			if mode in ( Gaffer.TweakPlug.Mode.Replace, Gaffer.TweakPlug.Mode.Create, Gaffer.TweakPlug.Mode.Remove ) :
 				continue
 
 			tweaks["tweaks"][0]["mode"].setValue( mode )
-			with self.assertRaisesRegex( RuntimeError, "Mode must be \"Replace\" when a previous connection exists" ) :
+			with self.assertRaisesRegex( RuntimeError, "mode is not compatible with an existing input connection" ) :
 				tweaks["out"].attributes( "/plane" )
 
 	def testPromoteTweaksPlug( self ) :
@@ -362,7 +392,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 			t["out"].attributes( "/light" )
 
 		inputShader = GafferSceneTest.TestShader()
-		badTweak["value"].setInput( inputShader["out"]["r"] )
+		badTweak["value"].setInput( inputShader["out"]["c"]["r"] )
 
 		with self.assertRaisesRegex( RuntimeError, "Cannot apply tweak \"badParameter\" because shader \"__shader\" does not have parameter \"badParameter\"" ) :
 			t["out"].attributes( "/light" )
@@ -375,7 +405,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		badTweak["name"].setValue( "badShader.p" )
 		self.assertEqual( t["out"].attributes( "/light" ), t["in"].attributes( "/light" ) )
 
-		badTweak["value"].setInput( inputShader["out"]["r"] )
+		badTweak["value"].setInput( inputShader["out"]["c"]["r"] )
 		badTweak["name"].setValue( "badParameter" )
 		self.assertEqual( t["out"].attributes( "/light" ), t["in"].attributes( "/light" ) )
 		self.assertEqual( t["out"].attributes( "/light" ), t["in"].attributes( "/light" ) )
@@ -488,16 +518,55 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		)
 		self.assertNotIn( "intensity", tweaks["out"].attributes( "/light" )["light"].outputShader().parameters )
 
+	def testRemoveExistingConnection( self ) :
+
+		# Build a network with a texture connected into the surface.
+
+		plane = GafferScene.Plane()
+
+		texture = GafferSceneTest.TestShader( "texture" )
+
+		surface = GafferSceneTest.TestShader( "surface" )
+		surface["parameters"]["c"].setInput( texture["out"]["c"] )
+		surface["type"].setValue( "test:surface" )
+
+		planeFilter = GafferScene.PathFilter()
+		planeFilter["paths"].setValue( IECore.StringVectorData( [ "/plane" ] ) )
+
+		assignment = GafferScene.ShaderAssignment()
+		assignment["in"].setInput( plane["out"] )
+		assignment["filter"].setInput( planeFilter["out"] )
+		assignment["shader"].setInput( surface["out"] )
+
+		originalNetwork = assignment["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( len( originalNetwork ), 2 )
+		self.assertEqual( originalNetwork.input( ( "surface", "c" ) ), ( "texture", "c" ) )
+
+		# Use a tweak to remove the connection and the value.
+
+		tweaks = GafferScene.ShaderTweaks()
+		tweaks["in"].setInput( assignment["out"] )
+		tweaks["filter"].setInput( planeFilter["out"] )
+		tweaks["shader"].setValue( "test:surface" )
+
+		tweaks["tweaks"].addChild( Gaffer.TweakPlug( "c", Gaffer.Color3fPlug() ) )
+		tweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Remove )
+
+		tweakedNetwork = tweaks["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( len( tweakedNetwork ), 1 )
+		self.assertFalse( tweakedNetwork.input( ( "surface", "c" ) ) )
+		self.assertNotIn( "c", tweakedNetwork.getShader( "surface" ).parameters )
+
 	def testWildcards( self ) :
 
 		shaderA_A = GafferSceneTest.TestShader( "A_A" )
 		shaderA_B = GafferSceneTest.TestShader( "A_B" )
-		shaderA_B["parameters"]["c"].setInput( shaderA_A["out"] )
+		shaderA_B["parameters"]["c"].setInput( shaderA_A["out"]["c"] )
 		shaderB_A = GafferSceneTest.TestShader( "B_A" )
-		shaderB_A["parameters"]["c"].setInput( shaderA_B["out"] )
+		shaderB_A["parameters"]["c"].setInput( shaderA_B["out"]["c"] )
 
 		light = GafferSceneTest.TestLight()
-		light["parameters"]["intensity"].setInput( shaderB_A["out"] )
+		light["parameters"]["intensity"].setInput( shaderB_A["out"]["c"] )
 
 		pathFilter = GafferScene.PathFilter()
 		pathFilter["paths"].setValue( IECore.StringVectorData( [ "/light" ] ) )
@@ -537,9 +606,9 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		self.assertEqual( tweaked.shaders()["B_A"].parameters["i"].value, 42 )
 
 		extraShader = GafferSceneTest.TestShader( "extra" )
-		tweaks["tweaks"][0]["value"].setInput( extraShader["out"]["r"] )
+		tweaks["tweaks"][0]["value"].setInput( extraShader["out"]["c"]["r"] )
 
-		with self.assertRaisesRegex( RuntimeError, r'Cannot apply tweak "\*_A.i" to "A_A.i" : Mode must be "Replace" when inserting a connection' ) :
+		with self.assertRaisesRegex( RuntimeError, r'Cannot apply tweak "\*_A.i" to "A_A.i" : Mode must be "Replace" or "Create" when inserting a connection' ) :
 			tweaks["out"].attributes( "/light" )
 
 		tweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Replace )
@@ -550,9 +619,8 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		# create two copies of the input network. Maybe in an ideal world, two separate tweaks using the same
 		# input network would also be shared?
 		self.assertEqual( len( tweaked ), 6 )
-		self.assertEqual( tweaked.inputConnections("A_A")[-1].source, IECoreScene.ShaderNetwork.Parameter( "extra", "out.r" ) )
-		self.assertEqual( tweaked.inputConnections("B_A")[-1].source, IECoreScene.ShaderNetwork.Parameter( "extra1", "out.r" ) )
-
+		self.assertEqual( tweaked.inputConnections("A_A")[-1].source, IECoreScene.ShaderNetwork.Parameter( "extra", "c.r" ) )
+		self.assertEqual( tweaked.inputConnections("B_A")[-1].source, IECoreScene.ShaderNetwork.Parameter( "extra1", "c.r" ) )
 
 		del tweaks["tweaks"][0]
 
@@ -560,7 +628,7 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 			Gaffer.TweakPlug( "*.c", imath.Color3f( 1, 2, 3), mode = Gaffer.TweakPlug.Mode.Add )
 		)
 
-		with self.assertRaisesRegex( RuntimeError, r'ShaderTweaks.out.attributes : Cannot apply tweak "\*.c" to "A_B.c" : Mode must be "Replace" when a previous connection exists' ) :
+		with self.assertRaisesRegex( RuntimeError, "mode is not compatible with an existing input connection" ) :
 			tweaks["out"].attributes( "/light" )
 
 		tweaks["tweaks"][0]["name"].setValue( "*.intensity" )
@@ -570,6 +638,75 @@ class ShaderTweaksTest( GafferSceneTest.SceneTestCase ) :
 		tweaked = tweaks["out"].attributes( "/light" )["light"]
 		self.assertEqual( len( tweaked ), 1 )
 		self.assertEqual( tweaked.shaders()["__shader"].parameters["intensity"].value, imath.Color3f( 1, 2, 3 ) )
+
+	def testClosureTweaks( self ) :
+
+		plane = GafferScene.Plane()
+
+		planeFilter = GafferScene.PathFilter()
+		planeFilter["paths"].setValue( IECore.StringVectorData( [ "/plane" ] ) )
+
+		surface = GafferSceneTest.TestShader( "surface" )
+		surface["type"].setValue( "test:surface" )
+		surface.loadShader( "mixClosures" )
+
+		shaderAssignment = GafferScene.ShaderAssignment()
+		shaderAssignment["in"].setInput( plane["out"] )
+		shaderAssignment["filter"].setInput( planeFilter["out"] )
+		shaderAssignment["shader"].setInput( surface["out"]["c"] )
+
+		shaderTweaks = GafferScene.ShaderTweaks()
+		shaderTweaks["in"].setInput( shaderAssignment["out"] )
+		shaderTweaks["filter"].setInput( planeFilter["out"] )
+		shaderTweaks["shader"].setValue( "test:surface" )
+
+		# Check we can create a new connection.
+
+		input = GafferSceneTest.TestShader( "input" )
+		input.loadShader( "mixClosures" )
+
+		shaderTweaks["tweaks"].addChild( Gaffer.TweakPlug( "a", GafferScene.ClosurePlug() ) )
+		shaderTweaks["tweaks"][0]["value"].setInput( input["out"]["c"] )
+		shaderTweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Create )
+
+		network = shaderTweaks["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( set( network.shaders().keys() ), { "surface", "input" } )
+		self.assertEqual( network.input( ( "surface", "a" ) ), ( "input", "c" ) )
+
+		# And that we can override an upstream connection.
+
+		upstreamInput = GafferSceneTest.TestShader( "upstreamInput" )
+		upstreamInput.loadShader( "mixClosures" )
+		surface["parameters"]["a"].setInput( upstreamInput["out"]["c"] )
+
+		network = shaderTweaks["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( set( network.shaders().keys() ), { "surface", "input" } )
+		self.assertEqual( network.input( ( "surface", "a" ) ), ( "input", "c" ) )
+
+		# And that if we have no input, the upstream connection
+		# is allowed to pass through.
+
+		shaderTweaks["tweaks"][0]["value"].setInput( None )
+
+		network = shaderTweaks["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( set( network.shaders().keys() ), { "surface", "upstreamInput" } )
+		self.assertEqual( network.input( ( "surface", "a" ) ), ( "upstreamInput", "c" ) )
+
+		# And that we can remove the upstream connection.
+
+		shaderTweaks["tweaks"][0]["mode"].setValue( Gaffer.TweakPlug.Mode.Remove )
+
+		network = shaderTweaks["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( set( network.shaders().keys() ), { "surface" } )
+		self.assertFalse( network.input( ( "surface", "a" ) ) )
+
+		# And that an input is ignored in remove mode, so we
+		# can still remove the upstream connection.
+
+		shaderTweaks["tweaks"][0]["value"].setInput( input["out"]["c"] )
+		network = shaderTweaks["out"].attributes( "/plane" )["test:surface"]
+		self.assertEqual( set( network.shaders().keys() ), { "surface" } )
+		self.assertFalse( network.input( ( "surface", "a" ) ) )
 
 if __name__ == "__main__":
 	unittest.main()